removeCategory

+++ backend/src/category-product/category-product.service.ts
@@ -1,3 +1,4 @@
+import { ProductModel } from 'src/product/product.model';
 import { CategoryProductDto } from './dto/category-product.dto';
 import { CategoryProductModel } from './category-product.model';
 import { Injectable, NotFoundException } from '@nestjs/common';
@@ -9,15 +10,17 @@ export class CategoryProductService {
   constructor(
     @InjectModel(CategoryProductModel)
     private readonly CategoryProductModel: ModelType<CategoryProductModel>,
+    @InjectModel(ProductModel)
+    private readonly ProductModel: ModelType<ProductModel>,
   ) {}
-  //создание типа товара
+  //создание категории товара
   async createCategoryProduct(dto: CategoryProductDto) {
     const categoryProduct = await this.CategoryProductModel.create(dto);
     if (!categoryProduct)
       throw new NotFoundException('Категория продукта не создан');
     return categoryProduct;
   }
-  // получение типов товаров
+  // получение категорий товаров
   async getCategoryProduct() {
     const categoryProduct = await this.CategoryProductModel.find()
       .populate('productType brand')
@@ -25,9 +28,13 @@ export class CategoryProductService {
     if (!categoryProduct) throw new NotFoundException('Категории не получены');
     return categoryProduct;
   }
-  // удаление типа товара
-  async removeProductType(id: string) {
-    //написать,когда будут продукты: делать проверку о наличии товаров с таким типом и только потом удальть,удалить тип из категории
+  // удаление категории товара
+  async removeCategoryProduct(id: string) {
+    //делаем запрос на товары, если товар с такой категорией существует, то не удаляем
+    const product = await this.ProductModel.findOne({ categoryId: id });
+
+    if (product)
+      return { message: 'Категория не удалёна,использутся в товарах' };
     const removeCategoryProduct =
       await this.CategoryProductModel.findByIdAndDelete(id).exec();
     if (!removeCategoryProduct)
